powsybl-config: move ClassicPlatformConfigProvider to another maven module and use a classpath only resolution mechanism
