+++ .gitignore
@@ -31,6 +31,7 @@
 /commons/target/
 /computation-local/target/
 /computation/target/
+/config-classic/target/
 /config-test/target/
 /contingency/target/
 /contingency/contingency-api/target/
